add navigation/StackNavigator file with all the stack.screens, import in App.js, then show login and register screens
